+++ CHANGELOG.md
@@ -1,15 +1,23 @@
-## 0.1.1-dev3
+## 0.1.2-dev2
 
 ### Enhancements
 
 * **Migrate Discord Source Connector to V2**
 
-## 0.1.1-dev1
+## 0.1.2-dev1
+
+### Enhancements
+
+* **Migrate Slack Source Connector to V2**
+* **Add Delta Table destination to v2**
+
+## 0.1.1
 
 ### Enhancements
 
 * **Update KDB.AI vectorstore integration to 1.4**
 * **Add sqlite and postgres source connectors**
+* **Add sampling functionality for indexers in fsspec connectors**
 
 ## 0.1.0